+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2020-04-20
+### Changed
+- Moved `solution` to its own file
+- Improved display of solutions
+
 ## [1.1.0] - 2020-04-18
 ### Added
 - useAuth, useInteraction, and useMission are hooks which make it super easy to interact with the database